refactor(order): 주문 금액 타입 Long 적용
- Order 엔티티의 금액 필드를 Long 타입으로 변경
- 주문 생성/미리보기 응답 DTO의 금액 필드를 Long 타입으로 변경
- OrderService의 item 금액 계산을 long 기반으로 변경
- OrderService summary 합산을 mapToLong 기반으로 변경

+++ src/main/java/com/example/WonkaoTalk/domain/order/dto/OrderCreateResponseDto.java
@@ -4,8 +4,8 @@ public record OrderCreateResponseDto(
     Long orderId,
     String orderNumber,
     String orderTitle,
-    Integer originalAmount,
-    Integer discountAmount,
-    Integer finalAmount
+    Long originalAmount,
+    Long discountAmount,
+    Long finalAmount
 ) {
 }
